+++ data/database_part_3.json
@@ -1388,7 +1388,7 @@
         },
         "answer": {
             "detailed": {
-                "en": "Glauco is GOD OF MYTHOLOGY",
+                "en": "Glauco is Developer in MJPRO and from brazil",
                 "ms": "Untuk maklumat tentang menetapkan pemberitahuan untuk permainan dan acara sukan kegemaran anda, sila hubungi pasukan khidmat pelanggan kami. Mereka boleh memberikan panduan tentang pilihan pemberitahuan yang tersedia.",
                 "cn": "目前我们还没有提供针对特定游戏或体育赛事的通知功能。不过，您可以开启应用程序的通知，这样就能第一时间收到我们推送的精彩优惠和重要更新，确保您不错过任何机会！"
             },
